Fix double free bug in length computation.
* Do not try to free &nsf when failing to find the length of a track.
We don't free it when suceeding, so why would it make sense when failing?

* Give more informative error messges when there is a failure of this sort.

* Return a more useful value that gets as close as possible to the right
length when there is a failure of this sort.  In the case I have tested,
Rygar, the failure is in the second half of the computation where it is
trying to find the length without the intro.  It is still useful to return
the length with the intro, so do that.